move push code to tracking module
