+++ Sources/Common/autogenerated/AutoMockable.generated.swift
@@ -80,58 +80,6 @@ import UserNotifications
  */
 
 /**
- Class to easily create a mocked version of the `DeviceAttributesHook` class.
- This class is equipped with functions and properties ready for you to mock!
-
- Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
- See the SDK documentation to learn the basics behind using the mock classes in the SDK.
- */
-public class DeviceAttributesHookMock: DeviceAttributesHook, Mock {
-    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
-    public var mockCalled: Bool = false //
-
-    public init() {
-        Mocks.shared.add(mock: self)
-    }
-
-    public func resetMock() {
-        customDeviceAttributesAddedCallsCount = 0
-        customDeviceAttributesAddedReceivedArguments = nil
-        customDeviceAttributesAddedReceivedInvocations = []
-
-        mockCalled = false // do last as resetting properties above can make this true
-    }
-
-    // MARK: - customDeviceAttributesAdded
-
-    /// Number of times the function was called.
-    public private(set) var customDeviceAttributesAddedCallsCount = 0
-    /// `true` if the function was ever called.
-    public var customDeviceAttributesAddedCalled: Bool {
-        customDeviceAttributesAddedCallsCount > 0
-    }
-
-    /// The arguments from the *last* time the function was called.
-    public private(set) var customDeviceAttributesAddedReceivedArguments: [String: Any]?
-    /// Arguments from *all* of the times that the function was called.
-    public private(set) var customDeviceAttributesAddedReceivedInvocations: [[String: Any]] = []
-    /**
-     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
-     */
-    public var customDeviceAttributesAddedClosure: (([String: Any]) -> Void)?
-
-    /// Mocked function for `customDeviceAttributesAdded(attributes: [String: Any])`. Your opportunity to return a
-    /// mocked value and check result of mock in test code.
-    public func customDeviceAttributesAdded(attributes: [String: Any]) {
-        mockCalled = true
-        customDeviceAttributesAddedCallsCount += 1
-        customDeviceAttributesAddedReceivedArguments = attributes
-        customDeviceAttributesAddedReceivedInvocations.append(attributes)
-        customDeviceAttributesAddedClosure?(attributes)
-    }
-}
-
-/**
  Class to easily create a mocked version of the `DeviceInfo` class.
  This class is equipped with functions and properties ready for you to mock!
 
@@ -904,42 +852,6 @@ public class HooksManagerMock: HooksManager, Mock {
      When setter of the property called, the value given to setter is set here.
      When the getter of the property called, the value set here will be returned. Your chance to mock the property.
      */
-    public var underlyingDeviceAttributesHooks: [DeviceAttributesHook] = []
-    /// `true` if the getter or setter of property is called at least once.
-    public var deviceAttributesHooksCalled: Bool {
-        deviceAttributesHooksGetCalled || deviceAttributesHooksSetCalled
-    }
-
-    /// `true` if the getter called on the property at least once.
-    public var deviceAttributesHooksGetCalled: Bool {
-        deviceAttributesHooksGetCallsCount > 0
-    }
-
-    public var deviceAttributesHooksGetCallsCount = 0
-    /// `true` if the setter called on the property at least once.
-    public var deviceAttributesHooksSetCalled: Bool {
-        deviceAttributesHooksSetCallsCount > 0
-    }
-
-    public var deviceAttributesHooksSetCallsCount = 0
-    /// The mocked property with a getter and setter.
-    public var deviceAttributesHooks: [DeviceAttributesHook] {
-        get {
-            mockCalled = true
-            deviceAttributesHooksGetCallsCount += 1
-            return underlyingDeviceAttributesHooks
-        }
-        set(value) {
-            mockCalled = true
-            deviceAttributesHooksSetCallsCount += 1
-            underlyingDeviceAttributesHooks = value
-        }
-    }
-
-    /**
-     When setter of the property called, the value given to setter is set here.
-     When the getter of the property called, the value set here will be returned. Your chance to mock the property.
-     */
     public var underlyingScreenViewHooks: [ScreenTrackingHook] = []
     /// `true` if the getter or setter of property is called at least once.
     public var screenViewHooksCalled: Bool {
@@ -977,8 +889,6 @@ public class HooksManagerMock: HooksManager, Mock {
         profileIdentifyHooksSetCallsCount = 0
         queueRunnerHooksGetCallsCount = 0
         queueRunnerHooksSetCallsCount = 0
-        deviceAttributesHooksGetCallsCount = 0
-        deviceAttributesHooksSetCallsCount = 0
         screenViewHooksGetCallsCount = 0
         screenViewHooksSetCallsCount = 0
         addCallsCount = 0
@@ -1543,42 +1453,6 @@ public class ModuleHookProviderMock: ModuleHookProvider, Mock {
      When setter of the property called, the value given to setter is set here.
      When the getter of the property called, the value set here will be returned. Your chance to mock the property.
      */
-    public var underlyingDeviceAttributesHook: DeviceAttributesHook?
-    /// `true` if the getter or setter of property is called at least once.
-    public var deviceAttributesHookCalled: Bool {
-        deviceAttributesHookGetCalled || deviceAttributesHookSetCalled
-    }
-
-    /// `true` if the getter called on the property at least once.
-    public var deviceAttributesHookGetCalled: Bool {
-        deviceAttributesHookGetCallsCount > 0
-    }
-
-    public var deviceAttributesHookGetCallsCount = 0
-    /// `true` if the setter called on the property at least once.
-    public var deviceAttributesHookSetCalled: Bool {
-        deviceAttributesHookSetCallsCount > 0
-    }
-
-    public var deviceAttributesHookSetCallsCount = 0
-    /// The mocked property with a getter and setter.
-    public var deviceAttributesHook: DeviceAttributesHook? {
-        get {
-            mockCalled = true
-            deviceAttributesHookGetCallsCount += 1
-            return underlyingDeviceAttributesHook
-        }
-        set(value) {
-            mockCalled = true
-            deviceAttributesHookSetCallsCount += 1
-            underlyingDeviceAttributesHook = value
-        }
-    }
-
-    /**
-     When setter of the property called, the value given to setter is set here.
-     When the getter of the property called, the value set here will be returned. Your chance to mock the property.
-     */
     public var underlyingScreenTrackingHook: ScreenTrackingHook?
     /// `true` if the getter or setter of property is called at least once.
     public var screenTrackingHookCalled: Bool {
@@ -1618,9 +1492,6 @@ public class ModuleHookProviderMock: ModuleHookProvider, Mock {
         queueRunnerHook = nil
         queueRunnerHookGetCallsCount = 0
         queueRunnerHookSetCallsCount = 0
-        deviceAttributesHook = nil
-        deviceAttributesHookGetCallsCount = 0
-        deviceAttributesHookSetCallsCount = 0
         screenTrackingHook = nil
         screenTrackingHookGetCallsCount = 0
         screenTrackingHookSetCallsCount = 0
