+++ Sources/Common/hooks/HooksManager.swift
@@ -4,7 +4,6 @@ public protocol HooksManager: AutoMockable {
     func add(key: HookModule, provider: ModuleHookProvider)
     var profileIdentifyHooks: [ProfileIdentifyHook] { get }
     var queueRunnerHooks: [QueueRunnerHook] { get }
-    var deviceAttributesHooks: [DeviceAttributesHook] { get }
     var screenViewHooks: [ScreenTrackingHook] { get }
 }
 
@@ -37,10 +36,6 @@ public class CioHooksManager: HooksManager {
 
     public var queueRunnerHooks: [QueueRunnerHook] {
         hookProviders.filter { $0.value.queueRunnerHook != nil }.map { $0.value.queueRunnerHook! }
-    }
-
-    public var deviceAttributesHooks: [DeviceAttributesHook] {
-        hookProviders.filter { $0.value.deviceAttributesHook != nil }.map { $0.value.deviceAttributesHook! }
     }
 
     public var screenViewHooks: [ScreenTrackingHook] {
