--- Sources/MessagingPush/Background Queue/Type/QueueTaskType.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-/// All the types of tasks the `MessagingPush` module runs in the background queue
-internal enum QueueTaskType: String {
-    case registerPushToken
-    case deletePushToken
-    case trackPushMetric
-}
