+++ Sources/MessagingPush/autogenerated/AutoMockable.generated.swift
@@ -82,58 +82,6 @@ import Common
  */
 
 /**
- Class to easily create a mocked version of the `DeviceAttributesProvider` class.
- This class is equipped with functions and properties ready for you to mock!
-
- Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
- See the SDK documentation to learn the basics behind using the mock classes in the SDK.
- */
-internal class DeviceAttributesProviderMock: DeviceAttributesProvider, Mock {
-    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
-    internal var mockCalled: Bool = false //
-
-    internal init() {
-        Mocks.shared.add(mock: self)
-    }
-
-    public func resetMock() {
-        getDefaultDeviceAttributesCallsCount = 0
-        getDefaultDeviceAttributesReceivedArguments = nil
-        getDefaultDeviceAttributesReceivedInvocations = []
-
-        mockCalled = false // do last as resetting properties above can make this true
-    }
-
-    // MARK: - getDefaultDeviceAttributes
-
-    /// Number of times the function was called.
-    internal private(set) var getDefaultDeviceAttributesCallsCount = 0
-    /// `true` if the function was ever called.
-    internal var getDefaultDeviceAttributesCalled: Bool {
-        getDefaultDeviceAttributesCallsCount > 0
-    }
-
-    /// The arguments from the *last* time the function was called.
-    internal private(set) var getDefaultDeviceAttributesReceivedArguments: (([String: Any]) -> Void)?
-    /// Arguments from *all* of the times that the function was called.
-    internal private(set) var getDefaultDeviceAttributesReceivedInvocations: [([String: Any]) -> Void] = []
-    /**
-     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
-     */
-    internal var getDefaultDeviceAttributesClosure: ((([String: Any]) -> Void) -> Void)?
-
-    /// Mocked function for `getDefaultDeviceAttributes(onComplete: @escaping ([String: Any]) -> Void)`. Your
-    /// opportunity to return a mocked value and check result of mock in test code.
-    internal func getDefaultDeviceAttributes(onComplete: @escaping ([String: Any]) -> Void) {
-        mockCalled = true
-        getDefaultDeviceAttributesCallsCount += 1
-        getDefaultDeviceAttributesReceivedArguments = onComplete
-        getDefaultDeviceAttributesReceivedInvocations.append(onComplete)
-        getDefaultDeviceAttributesClosure?(onComplete)
-    }
-}
-
-/**
  Class to easily create a mocked version of the `MessagingPushInstance` class.
  This class is equipped with functions and properties ready for you to mock!
