+++ Sources/Tracking/Util/SdkInitializedUtil.swift
@@ -2,6 +2,7 @@ import Common
 import Foundation
 
 public protocol SdkInitializedUtil: AutoMockable {
+    var customerio: CustomerIO? { get }
     var isInitlaized: Bool { get }
     var postInitializedData: (siteId: String, diGraph: DIGraph)? { get }
 }
@@ -12,6 +13,12 @@ public class SdkInitializedUtilImpl: SdkInitializedUtil {
     // Try to not use dependencies in this class as it's contructed before
     // DI graph could be populated. So, production code calls this constructor.
     public init() {}
+
+    public var customerio: CustomerIO? {
+        guard isInitlaized else { return nil }
+
+        return CustomerIO.shared
+    }
 
     public var isInitlaized: Bool {
         CustomerIO.shared.diGraph != nil
