+++ Sources/Tracking/autogenerated/AutoMockable.generated.swift
@@ -285,6 +285,19 @@ public class CustomerIOInstanceMock: CustomerIOInstance, Mock {
         screenEncodableReceivedInvocations = []
 
         mockCalled = false // do last as resetting properties above can make this true
+        registerDeviceTokenCallsCount = 0
+        registerDeviceTokenReceivedArguments = nil
+        registerDeviceTokenReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+        deleteDeviceTokenCallsCount = 0
+
+        mockCalled = false // do last as resetting properties above can make this true
+        trackMetricCallsCount = 0
+        trackMetricReceivedArguments = nil
+        trackMetricReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
     }
 
     // MARK: - identify
@@ -455,6 +468,137 @@ public class CustomerIOInstanceMock: CustomerIOInstance, Mock {
         screenEncodableReceivedInvocations.append((name: name, data: AnyEncodable(data)))
         screenEncodableClosure?(name, AnyEncodable(data))
     }
+
+    // MARK: - registerDeviceToken
+
+    /// Number of times the function was called.
+    public private(set) var registerDeviceTokenCallsCount = 0
+    /// `true` if the function was ever called.
+    public var registerDeviceTokenCalled: Bool {
+        registerDeviceTokenCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    public private(set) var registerDeviceTokenReceivedArguments: String?
+    /// Arguments from *all* of the times that the function was called.
+    public private(set) var registerDeviceTokenReceivedInvocations: [String] = []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    public var registerDeviceTokenClosure: ((String) -> Void)?
+
+    /// Mocked function for `registerDeviceToken(_ deviceToken: String)`. Your opportunity to return a mocked value and
+    /// check result of mock in test code.
+    public func registerDeviceToken(_ deviceToken: String) {
+        mockCalled = true
+        registerDeviceTokenCallsCount += 1
+        registerDeviceTokenReceivedArguments = deviceToken
+        registerDeviceTokenReceivedInvocations.append(deviceToken)
+        registerDeviceTokenClosure?(deviceToken)
+    }
+
+    // MARK: - deleteDeviceToken
+
+    /// Number of times the function was called.
+    public private(set) var deleteDeviceTokenCallsCount = 0
+    /// `true` if the function was ever called.
+    public var deleteDeviceTokenCalled: Bool {
+        deleteDeviceTokenCallsCount > 0
+    }
+
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    public var deleteDeviceTokenClosure: (() -> Void)?
+
+    /// Mocked function for `deleteDeviceToken()`. Your opportunity to return a mocked value and check result of mock in
+    /// test code.
+    public func deleteDeviceToken() {
+        mockCalled = true
+        deleteDeviceTokenCallsCount += 1
+        deleteDeviceTokenClosure?()
+    }
+
+    // MARK: - trackMetric
+
+    /// Number of times the function was called.
+    public private(set) var trackMetricCallsCount = 0
+    /// `true` if the function was ever called.
+    public var trackMetricCalled: Bool {
+        trackMetricCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    public private(set) var trackMetricReceivedArguments: (deliveryID: String, event: Metric, deviceToken: String)?
+    /// Arguments from *all* of the times that the function was called.
+    public private(set) var trackMetricReceivedInvocations: [(deliveryID: String, event: Metric, deviceToken: String)] =
+        []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    public var trackMetricClosure: ((String, Metric, String) -> Void)?
+
+    /// Mocked function for `trackMetric(deliveryID: String, event: Metric, deviceToken: String)`. Your opportunity to
+    /// return a mocked value and check result of mock in test code.
+    public func trackMetric(deliveryID: String, event: Metric, deviceToken: String) {
+        mockCalled = true
+        trackMetricCallsCount += 1
+        trackMetricReceivedArguments = (deliveryID: deliveryID, event: event, deviceToken: deviceToken)
+        trackMetricReceivedInvocations.append((deliveryID: deliveryID, event: event, deviceToken: deviceToken))
+        trackMetricClosure?(deliveryID, event, deviceToken)
+    }
+}
+
+/**
+ Class to easily create a mocked version of the `DeviceAttributesProvider` class.
+ This class is equipped with functions and properties ready for you to mock!
+
+ Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
+ See the SDK documentation to learn the basics behind using the mock classes in the SDK.
+ */
+internal class DeviceAttributesProviderMock: DeviceAttributesProvider, Mock {
+    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
+    internal var mockCalled: Bool = false //
+
+    internal init() {
+        Mocks.shared.add(mock: self)
+    }
+
+    public func resetMock() {
+        getDefaultDeviceAttributesCallsCount = 0
+        getDefaultDeviceAttributesReceivedArguments = nil
+        getDefaultDeviceAttributesReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+    }
+
+    // MARK: - getDefaultDeviceAttributes
+
+    /// Number of times the function was called.
+    internal private(set) var getDefaultDeviceAttributesCallsCount = 0
+    /// `true` if the function was ever called.
+    internal var getDefaultDeviceAttributesCalled: Bool {
+        getDefaultDeviceAttributesCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    internal private(set) var getDefaultDeviceAttributesReceivedArguments: (([String: Any]) -> Void)?
+    /// Arguments from *all* of the times that the function was called.
+    internal private(set) var getDefaultDeviceAttributesReceivedInvocations: [([String: Any]) -> Void] = []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    internal var getDefaultDeviceAttributesClosure: ((([String: Any]) -> Void) -> Void)?
+
+    /// Mocked function for `getDefaultDeviceAttributes(onComplete: @escaping ([String: Any]) -> Void)`. Your
+    /// opportunity to return a mocked value and check result of mock in test code.
+    internal func getDefaultDeviceAttributes(onComplete: @escaping ([String: Any]) -> Void) {
+        mockCalled = true
+        getDefaultDeviceAttributesCallsCount += 1
+        getDefaultDeviceAttributesReceivedArguments = onComplete
+        getDefaultDeviceAttributesReceivedInvocations.append(onComplete)
+        getDefaultDeviceAttributesClosure?(onComplete)
+    }
 }
 
 /**
@@ -470,6 +614,42 @@ public class SdkInitializedUtilMock: SdkInitializedUtil, Mock {
 
     public init() {
         Mocks.shared.add(mock: self)
+    }
+
+    /**
+     When setter of the property called, the value given to setter is set here.
+     When the getter of the property called, the value set here will be returned. Your chance to mock the property.
+     */
+    public var underlyingCustomerio: CustomerIO?
+    /// `true` if the getter or setter of property is called at least once.
+    public var customerioCalled: Bool {
+        customerioGetCalled || customerioSetCalled
+    }
+
+    /// `true` if the getter called on the property at least once.
+    public var customerioGetCalled: Bool {
+        customerioGetCallsCount > 0
+    }
+
+    public var customerioGetCallsCount = 0
+    /// `true` if the setter called on the property at least once.
+    public var customerioSetCalled: Bool {
+        customerioSetCallsCount > 0
+    }
+
+    public var customerioSetCallsCount = 0
+    /// The mocked property with a getter and setter.
+    public var customerio: CustomerIO? {
+        get {
+            mockCalled = true
+            customerioGetCallsCount += 1
+            return underlyingCustomerio
+        }
+        set(value) {
+            mockCalled = true
+            customerioSetCallsCount += 1
+            underlyingCustomerio = value
+        }
     }
 
     /**
@@ -545,6 +725,9 @@ public class SdkInitializedUtilMock: SdkInitializedUtil, Mock {
     }
 
     public func resetMock() {
+        customerio = nil
+        customerioGetCallsCount = 0
+        customerioSetCallsCount = 0
         isInitlaizedGetCallsCount = 0
         isInitlaizedSetCallsCount = 0
         postInitializedData = nil
